생성 - pie graph 를 그리는데 필요한 구조체들을 별도 파일로 이동
